Improves errors

+++ cmd/destroy.go
@@ -12,13 +12,17 @@ var destroyCmd = &cobra.Command{
 	Short: "Destroys all AWS resources created by the apply/create",
 	Long: `This command runs terraform workspace select, terraform destroy and terraform workspace delete`,
 	Run: func(cmd *cobra.Command, args []string) {
-		configFile := ReadFile(WEBSOCKET_CONFIG_FILE_PATH)
+		configFile, readFileError := ReadFile(WEBSOCKET_CONFIG_FILE_PATH)
 
 		if len(configFile) == 0 {
 			errorMessage.Println(CONFIG_FILE_NOT_FOUND_MESSAGE)
 			return
 		}
 
+		if readFileError != nil {
+			errorMessage.Println(readFileError)
+			return
+		}
 
 		argumentsValid, argumentInvalidMessage, argumentsForMessage := checkForValidArguments("terraform destroy", args, configFile)
 
@@ -29,7 +33,7 @@ var destroyCmd = &cobra.Command{
 
 		currentDirectory, error := os.Getwd()
 		if error != nil {
-			errorMessage.Println("ERROR: The current directory path was not retrieved: %v", error)
+			errorMessage.Println(CURRENT_DIRECTORY_ERROR_MESSAGE, error)
 			return
 		}
